requirements.txt
google-generativeai
fastapi
pydantic
uvicorn
sqlalchemy
pyarrow
pandas
